// src/lib/models/ConfigPoint2D.ts
export class ConfigPoint2D {
    
    innerColor: string = "black";
    borderColor: string = "green";         
    color: string = "red";
    position: string = "";
    borderWidth: number = 1;
    viewName:boolean = false;
    enabled: boolean = true;
    radial: boolean = false;
    message: string = "";
    closed: any;
    selectable: boolean = true;
    visible: boolean = true;

}